moved parsePrice logic;
